Bugfix in examples/rl/environments/shipsteer/shipbench_pgpe.py

+++ examples/rl/environments/shipsteer/shipbench_pgpe.py
@@ -20,7 +20,7 @@ from pybrain.rl.experiments import EpisodicExperiment
 
 batch=1 #number of samples per learning step
 prnts=50 #number of learning steps after results are printed
-epis=2000/batch/prnts #number of roleouts
+epis=int(2000/batch/prnts) #number of roleouts
 numbExp=10 #number of experiments
 et = ExTools(batch, prnts) #tool for printing and plotting
